Fix several issues commented
