+++ rl_algorithms/common/networks/brain.py
@@ -9,12 +9,9 @@
 import torch
 import torch.nn as nn
 
-from rl_algorithms.common.helper_functions import (
-    identity,
-    infer_leading_dims,
-    restore_leading_dims,
-)
+from rl_algorithms.common.helper_functions import identity
 from rl_algorithms.dqn.networks import IQNMLP
+from rl_algorithms.recurrent.utils import infer_leading_dims, restore_leading_dims
 from rl_algorithms.registry import build_backbone, build_head
 from rl_algorithms.utils.config import ConfigDict
 
@@ -99,25 +96,36 @@ class GRUBrain(Brain):
         prev_action: torch.Tensor,
         prev_reward: torch.Tensor,
     ):
-        """Forward method implementation. Use in get_action method in agent."""
+        """Forward method implementation. Use in get_action method in agent.
+        We adopted RL^2 algorithm's architecture, which recieve
+        previous action & rewrad as RNN input.
+
+        RL^2 paper : https://arxiv.org/pdf/1611.02779.pdf
+
+        Args:
+            x (torch.Tensor): state of the transition
+            hidden (torch.Tensor): hidden_state of the transition
+            prev_action (torch.Tensor): previous transition's action
+            prev_reward (torch.Tensor): previous transition's reward
+        """
         if isinstance(self.backbone, nn.Module):
             x = x / 255.0
-            lead_dim, T, B, x_shape = infer_leading_dims(x, 3)
+            lead_dim, batch_len, seq_len, x_shape = infer_leading_dims(x, 3)
 
             backbone_out = self.backbone(
-                x.view(T * B, *x_shape)
+                x.view(batch_len * seq_len, *x_shape)
             )  # Fold if T dimension.
         else:
             if len(x.shape) == 1:
                 x = x.reshape(1, 1, -1)
-            lead_dim, T, B, x_shape = infer_leading_dims(x, 1)
+            lead_dim, batch_len, seq_len, x_shape = infer_leading_dims(x, 1)
             backbone_out = x
         lstm_input = self.fc(backbone_out)
         lstm_input = torch.cat(
             [
-                lstm_input.view(T, B, -1),
-                prev_action.view(T, B, -1),
-                prev_reward.view(T, B, 1),
+                lstm_input.view(batch_len, seq_len, -1),
+                prev_action.view(batch_len, seq_len, -1),
+                prev_reward.view(batch_len, seq_len, 1),
             ],
             dim=2,
         )
@@ -125,10 +133,10 @@ class GRUBrain(Brain):
         hidden = None if hidden is None else hidden
         lstm_out, hidden = self.gru(lstm_input, hidden)
 
-        q = self.head(lstm_out.contiguous().view(T * B, -1))
+        q = self.head(lstm_out.contiguous().view(batch_len * seq_len, -1))
 
         # Restore leading dimensions: [T,B], [B], or [], as input.
-        q = restore_leading_dims(q, lead_dim, T, B)
+        q = restore_leading_dims(q, lead_dim, batch_len, seq_len)
 
         return q, hidden
 
@@ -143,22 +151,22 @@ class GRUBrain(Brain):
         """Get output value for calculating loss."""
         if isinstance(self.backbone, nn.Module):
             x = x / 255.0
-            lead_dim, T, B, x_shape = infer_leading_dims(x, 3)
+            lead_dim, batch_len, seq_len, x_shape = infer_leading_dims(x, 3)
 
             backbone_out = self.backbone(
-                x.view(T * B, *x_shape)
+                x.view(batch_len * seq_len, *x_shape)
             )  # Fold if T dimension.
         else:
             if len(x.shape) == 1:
                 x = x.reshape(1, 1, -1)
-            lead_dim, T, B, x_shape = infer_leading_dims(x, 1)
+            lead_dim, batch_len, seq_len, x_shape = infer_leading_dims(x, 1)
             backbone_out = x
         lstm_input = self.fc(backbone_out)
         lstm_input = torch.cat(
             [
-                lstm_input.view(T, B, -1),
-                prev_action.view(T, B, -1),
-                prev_reward.view(T, B, 1),
+                lstm_input.view(batch_len, seq_len, -1),
+                prev_action.view(batch_len, seq_len, -1),
+                prev_reward.view(batch_len, seq_len, 1),
             ],
             dim=2,
         )
@@ -168,19 +176,23 @@ class GRUBrain(Brain):
 
         if isinstance(self.head, IQNMLP):
             quantile_values, quantiles = self.head.forward_(
-                lstm_out.contiguous().view(T * B, -1), n_tau_samples
+                lstm_out.contiguous().view(batch_len * seq_len, -1), n_tau_samples
             )
             # Restore leading dimensions: [T,B], [B], or [], as input.
             quantile_values = restore_leading_dims(
-                quantile_values, lead_dim, T * n_tau_samples, B
+                quantile_values, lead_dim, batch_len * n_tau_samples, seq_len
             )
-            quantiles = restore_leading_dims(quantiles, lead_dim, T * n_tau_samples, B)
+            quantiles = restore_leading_dims(
+                quantiles, lead_dim, batch_len * n_tau_samples, seq_len
+            )
 
             return quantile_values, quantiles, hidden
         else:
-            head_out = self.head.forward_(lstm_out.contiguous().view(T * B, -1))
+            head_out = self.head.forward_(
+                lstm_out.contiguous().view(batch_len * seq_len, -1)
+            )
             if len(head_out) != 2:  # C51 output is not going to be restore.
-                head_out = restore_leading_dims(head_out, lead_dim, T, B)
+                head_out = restore_leading_dims(head_out, lead_dim, batch_len, seq_len)
             return head_out, hidden
 
     def calculate_fc_input_size(self, state_dim: tuple):
